+++ backend/de.metas.adempiere.adempiere/base/src/main/java/de/metas/organization/LocalDateAndOrgId.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.NonNull;
 import lombok.ToString;
 
+import javax.annotation.Nullable;
 import java.sql.Timestamp;
 import java.time.Instant;
 import java.time.LocalDate;
@@ -14,7 +15,7 @@ import java.util.function.Function;
 
 @EqualsAndHashCode(doNotUseGetters = true)
 @ToString(doNotUseGetters = true)
-public class LocalDateAndOrgId
+public class LocalDateAndOrgId implements Comparable<LocalDateAndOrgId>
 {
 	@NonNull private final LocalDate localDate;
 	@Getter @NonNull private final OrgId orgId;
@@ -52,4 +53,26 @@ public class LocalDateAndOrgId
 	{
 		return Timestamp.from(toInstant(orgMapper));
 	}
+
+	@Override
+	public int compareTo(final @Nullable LocalDateAndOrgId o)
+	{
+		return compareToByLocalDate(o);
+	}
+
+	/**
+	 * In {@code LocalDateAndOrgId}, only the localDate is the actual data, while orgId is used to give context for reading & writing purposes.
+	 * A calendar date is directly comparable to another one, without regard of "from which org has this date been extracted?"
+	 * That's why a comparison by local date is enough to provide correct ordering, even with different {@code OrgId}s.
+	 *
+	 * @see #compareTo(LocalDateAndOrgId)
+	 */
+	private int compareToByLocalDate(final @Nullable LocalDateAndOrgId o)
+	{
+		if (o == null)
+		{
+			return 1;
+		}
+		return this.localDate.compareTo(o.localDate);
+}
 }