+++ backend/de.metas.swat/de.metas.swat.base/src/main/java/de/metas/order/inoutcandidate/OrderLineShipmentScheduleHandler.java
@@ -174,6 +174,8 @@ public class OrderLineShipmentScheduleHandler extends ShipmentScheduleHandler
 		shipmentSchedule.setAD_Table_ID(getTableId(I_C_OrderLine.class));
 		shipmentSchedule.setRecord_ID(orderLine.getC_OrderLine_ID());
 
+		shipmentSchedule.setM_SectionCode(orderLine.getM_SectionCode());
+
 		shipmentSchedule.setQtyReserved(BigDecimal.ZERO.max(orderLine.getQtyReserved())); // task 09358: making sure that negative qtyOrdered are not propagated to the shipment sched
 
 		// 08255 : initialize the qty order calculated