+++ misc/services/mobile-webui/mobile-webui-frontend/src/components/dialogs/GetQuantityDialog.jsx
@@ -8,8 +8,10 @@ import QtyInputField from '../QtyInputField';
 import QtyReasonsRadioGroup from '../QtyReasonsRadioGroup';
 import * as ws from '../../utils/websocket';
 import { qtyInfos } from '../../utils/qtyInfos';
+import { formatQtyToHumanReadable } from '../../utils/qtys';
 
 const GetQuantityDialog = ({
+  userInfo,
   qtyInitial,
   qtyTarget,
   qtyCaption,
@@ -28,9 +30,9 @@ const GetQuantityDialog = ({
   const onQtyEntered = (qtyInfo) => setQtyInfo(qtyInfo);
   const onReasonSelected = (reason) => setRejectedReason(reason);
 
-  const requiredQtyRejectedReason = Array.isArray(qtyRejectedReasons) && qtyRejectedReasons.length > 0;
+  const isQtyRejectedRequired = Array.isArray(qtyRejectedReasons) && qtyRejectedReasons.length > 0;
   const qtyRejected =
-    requiredQtyRejectedReason && qtyInfos.isValid(qtyInfo)
+    isQtyRejectedRequired && qtyInfos.isValid(qtyInfo)
       ? Math.max(qtyTarget - qtyInfos.toNumberOrString(qtyInfo), 0)
       : 0;
 
@@ -40,6 +42,7 @@ const GetQuantityDialog = ({
     if (allValid) {
       onQtyChange({
         qtyEnteredAndValidated: qtyInfos.toNumberOrString(qtyInfo),
+        qtyRejected,
         qtyRejectedReason: qtyRejected > 0 ? rejectedReason : null,
       });
     }
@@ -77,12 +80,19 @@ const GetQuantityDialog = ({
           <div className="message-body">
             <table className="table">
               <tbody>
-                <tr>
-                  <th>{qtyCaption}</th>
-                  <td>
-                    {qtyTarget > 0 ? qtyTarget : 0} {uom}
-                  </td>
-                </tr>
+                {qtyCaption && (
+                  <tr>
+                    <th>{qtyCaption}</th>
+                    <td>{formatQtyToHumanReadable({ qty: Math.max(qtyTarget, 0), uom })}</td>
+                  </tr>
+                )}
+                {userInfo &&
+                  userInfo.map((item) => (
+                    <tr key={`userInfo_${item.caption}`}>
+                      <th>{item.caption}</th>
+                      <td>{item.value}</td>
+                    </tr>
+                  ))}
                 <tr>
                   <th>Qty</th>
                   <td>
@@ -116,25 +126,23 @@ const GetQuantityDialog = ({
                     </td>
                   </tr>
                 )}
-                {requiredQtyRejectedReason && (
-                  <tr>
-                    <th>{trl('general.QtyRejected')}</th>
-                    <td>
-                      {qtyRejected} {uom}
-                    </td>
-                  </tr>
-                )}
-                {requiredQtyRejectedReason && (
-                  <tr>
-                    <td colSpan={2}>
-                      <QtyReasonsRadioGroup
-                        reasons={qtyRejectedReasons}
-                        selectedReason={rejectedReason}
-                        disabled={qtyRejected === 0}
-                        onReasonSelected={onReasonSelected}
-                      />
-                    </td>
-                  </tr>
+                {qtyRejected > 0 && (
+                  <>
+                    <tr>
+                      <th>{trl('general.QtyRejected')}</th>
+                      <td>{formatQtyToHumanReadable({ qty: qtyRejected, uom })}</td>
+                    </tr>
+                    <tr>
+                      <td colSpan={2}>
+                        <QtyReasonsRadioGroup
+                          reasons={qtyRejectedReasons}
+                          selectedReason={rejectedReason}
+                          disabled={qtyRejected === 0}
+                          onReasonSelected={onReasonSelected}
+                        />
+                      </td>
+                    </tr>
+                  </>
                 )}
               </tbody>
             </table>
@@ -156,9 +164,10 @@ const GetQuantityDialog = ({
 
 GetQuantityDialog.propTypes = {
   // Properties
+  userInfo: PropTypes.array,
   qtyInitial: PropTypes.number,
   qtyTarget: PropTypes.number.isRequired,
-  qtyCaption: PropTypes.string.isRequired,
+  qtyCaption: PropTypes.string,
   uom: PropTypes.string.isRequired,
   qtyRejectedReasons: PropTypes.arrayOf(PropTypes.object),
   scaleDevice: PropTypes.object,